вынесение компонентов
