+++ index.html
@@ -14,7 +14,7 @@
     <meta name="twitter:card" content="summary_large_image">
 
     <title>V-Boost</title>
-    <link rel="icon" type="image/png" href="/src/assets/icons/Logo.svg" />
+    <link rel="icon" type="image/png" href="/src/assets/icons/favicon.svg"/>
   </head>
   <body>
     <div id="root" style="height: 100%;"></div>
