+++ src/assets/icons/SearchIcon.jsx
@@ -1,4 +1,4 @@
-const Search = () => {
+const SearchIcon = () => {
     return (
         <svg width="25" height="25" viewBox="0 0 25 25" fill="none" xmlns="http://www.w3.org/2000/svg">
             <path
@@ -8,4 +8,4 @@ const Search = () => {
     )
 }
 
-export default Search;
+export default SearchIcon;
