+++ src/assets/icons/index.js
@@ -1,4 +1,11 @@
-export * from './ApplePay.jsx';
-export * from './PayPal.jsx';
-export * from './Bitcoin.jsx';
-export * from './Games.jsx';
+export {default as ApplePay} from './ApplePay.jsx';
+export {default as PayPal} from './PayPal.jsx';
+export {default as Bitcoin} from './Bitcoin.jsx';
+export {default as GogglePay} from './GogglePay.jsx';
+export {default as Visa} from './Visa.jsx';
+export {default as Mastercard} from './Mastercard.jsx';
+export {default as YouTube} from './YouTube.jsx';
+export {default as Telegram} from './Telegram.jsx';
+export {default as Mail} from './Mail.jsx';
+export {default as WhatsApp} from './WhatsApp.jsx';
+export {default as Games} from './Games.jsx';
