+++ src/pages/BoosterMainPage.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Header from "../layouts/common/Header"
+import Header from "../layouts/common/header/Header.jsx"
 import { Outlet } from "react-router";
