+++ src/pages/HomePage.jsx
@@ -1,7 +1,7 @@
-import Header from '../layouts/common/Header.jsx';
+import Header from '../layouts/common/header/Header.jsx';
 import { Outlet } from "react-router";
 import '../styles/HomePage.css';
-import Footer from '../layouts/common/Footer.jsx';
+import Footer from '../layouts/common/Footer/Footer.jsx';
 import { Box } from '@mui/material';
 
 const HomePage = () => {
